@@ -3,6 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "type": "module",
+  "main": "server.js",
   "scripts": {
     "start": "cross-env NODE_ENV=production node ./server.js",
     "start:dev": "cross-env NODE_ENV=development nodemon ./server.js",
@@ -14,11 +15,12 @@
     "cross-env": "7.0.3",
     "express": "4.17.1",
     "joi": "^17.10.1",
+    "joi-phone-number": "^5.1.1",
     "morgan": "1.10.0",
     "nanoid": "^4.0.2"
   },
   "devDependencies": {
-    "eslint": "7.19.0",
+    "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-config-standard": "16.0.3",
     "eslint-plugin-import": "2.25.3",
