@@ -1,25 +1,26 @@
-const express = require('express')
-const logger = require('morgan')
-const cors = require('cors')
+import express from 'express';
+import cors from 'cors';
+import logger from 'morgan';
 
-const contactsRouter = require('./routes/api/contacts')
 
-const app = express()
+import contactsRouter from './routes/api/contacts.js';
 
-const formatsLogger = app.get('env') === 'development' ? 'dev' : 'short'
+const app = express();
 
-app.use(logger(formatsLogger))
-app.use(cors())
-app.use(express.json())
+const formatsLogger = app.get('env') === 'development' ? 'dev' : 'short';
 
-app.use('/api/contacts', contactsRouter)
+app.use(logger(formatsLogger));
+app.use(cors());
+app.use(express.json());
+
+app.use('/api/contacts', contactsRouter);
 
 app.use((req, res) => {
-  res.status(404).json({ message: 'Not found' })
-})
+  res.status(404).json({ message: 'Not found' });
+});
 
-app.use((err, req, res, next) => {
-  res.status(500).json({ message: err.message })
-})
+app.use((err, req, res) => {
+  res.status(500).json({ message: err.message });
+});
 
-module.exports = app
+export default app;
