@@ -1,18 +1,55 @@
-// const fs = require('fs/promises')
+import fs from 'fs/promises';
+import path from 'path';
+import { nanoid } from 'nanoid';
 
-const listContacts = async () => {}
+const contactsPath = path.resolve('models', 'contacts.json');
 
-const getContactById = async (contactId) => {}
+const listContacts = async () => {
+  const data = await fs.readFile(contactsPath)
+  return JSON.parse(data)
+};
 
-const removeContact = async (contactId) => {}
+const getById = async (id) => {
+  const allContacts = await listContacts();
+  const searchedContact = allContacts.find(item => item.id === id);
+  return searchedContact || null;
+}
 
-const addContact = async (body) => {}
+const removeContact = async (id) => {
+  const allContacts = await listContacts();
+  const index = allContacts.findIndex(item => item.id === id);
+  if (index === -1) {
+    return null;
+  }
+  const [result] = allContacts.splice(index, 1);
+  await fs.writeFile(contactsPath, JSON.stringify(allContacts, null, 2));
+  return result;
+}
 
-const updateContact = async (contactId, body) => {}
+const addContact = async (body) => {
+  const { name, email, phone } = body
+  const allContacts = await listContacts();
+  const newContact = { id: nanoid(), name, email, phone };
+  allContacts.push(newContact);
+  await fs.writeFile(contactsPath, JSON.stringify(allContacts, null, 2));
+  return newContact || null;
+}
 
-module.exports = {
+const updateContact = async (id, toUpdate) => {
+  const allContacts = await listContacts();
+  const index = allContacts.findIndex(item => item.id === id);
+  if (index === -1) {
+    return null;
+  }
+  const updatedContact = Object.assign(allContacts[index], toUpdate);
+  allContacts.splice(index, 1, updatedContact);
+  await fs.writeFile(contactsPath, JSON.stringify(allContacts, null, 2));
+  return updatedContact;
+}
+
+export default {
   listContacts,
-  getContactById,
+  getById,
   removeContact,
   addContact,
   updateContact,
