@@ -1,8 +1,8 @@
 module.exports = {
   env: {
-    commonjs: true,
-    es2021: true,
-    node: true,
+    "commonjs": true,
+    "es2021": true,
+    "node": true,
   },
   extends: ['standard', 'prettier'],
   parserOptions: {
